Extract common expression shims to 'shims', make SetF work (die) on django < 1.8
