+++ django_mysql/models/expressions.py
@@ -1,5 +1,9 @@
-from django.db.models.expressions import BaseExpression, F, Value
+# -*- coding:utf-8 -*-
+from __future__ import unicode_literals
 
+from django.db.models import F
+
+from django_mysql.shims.expressions import BaseExpression, Value
 from django_mysql.utils import collapse_spaces
